Drop the `Feature` suffix from Reducers
