+++ FoodSpec.xcodeproj/xcuserdata/victorsocaciu.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -49,12 +49,12 @@
 		<key>FoodSpec-Release.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>2</integer>
 		</dict>
 		<key>FoodSpec.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>3</integer>
+			<integer>5</integer>
 		</dict>
 		<key>MyPlayground (Playground) 1.xcscheme</key>
 		<dict>
