+++ food-spec/Sources/FoodDetails/FoodDetails.swift
@@ -1,149 +1,47 @@
-import SwiftUI
+import Foundation
 import ComposableArchitecture
 import Shared
 import QuantityPicker
 
-public struct FoodDetails: View {
-    @Bindable var store: StoreOf<FoodDetailsFeature>
+@Reducer
+public struct FoodDetails {
+    @ObservableState
+    public struct State: Hashable {
+        let originalFood: Food
+        var food: Food
+        var quantityPicker: QuantityPicker.State = .init()
 
-    public init(store: StoreOf<FoodDetailsFeature>) {
-        self.store = store
-    }
-
-    public var body: some View {
-        ScrollView {
-            Section {
-                quantityPicker
-                energy
-                protein
-                carbohydrates
-                fat
-                cholesterol
-                potassium
-                sodium
-                energyBreakdown
+        public init(food: Food, quantity: Quantity? = nil) {
+            self.originalFood = food
+            self.food = food
+            if let quantity {
+                self.quantityPicker = .init(quantity: quantity)
+                self.food = food.changingServingSize(to: quantity)
             }
-            .padding(.horizontal)
-        }
-        .scrollDismissesKeyboard(.immediately)
-        .navigationTitle(store.food.name.capitalized)
-    }
-
-    var energy: some View {
-        NutritionalValueCard(
-            model: .init(
-                title: "Energy",
-                value: store.food.energy.measurement,
-                breakdown: []
-            )
-        )
-    }
-
-    var protein: some View {
-        NutritionalValueCard(
-            model: .init(
-                title: "Protein",
-                value: store.food.protein.measurement,
-                breakdown: []
-            )
-        )
-    }
-
-    var carbohydrates: some View {
-        NutritionalValueCard(
-            model: .init(
-                title: "Carbohydrate",
-                value: store.food.carbohydrate.measurement,
-                breakdown: [
-                    .init(
-                        title: "Fiber",
-                        value: store.food.fiber.measurement
-                    ),
-                    .init(
-                        title: "Sugar",
-                        value: store.food.sugar.measurement
-                    ),
-                ]
-            )
-        )
-    }
-
-    var fat: some View {
-        NutritionalValueCard(
-            model: .init(
-                title: "Total Fat",
-                value: store.food.fatTotal.measurement,
-                breakdown: [
-                    .init(
-                        title: "Saturated Fat",
-                        value: store.food.fatSaturated.measurement
-                    )
-                ]
-            )
-        )
-    }
-
-    var cholesterol: some View {
-        NutritionalValueCard(
-            model: .init(
-                title: "Cholesterol",
-                value: store.food.cholesterol.measurement.converted(to: .milligrams),
-                breakdown: []
-            )
-        )
-    }
-
-    var potassium: some View {
-        NutritionalValueCard(
-            model: .init(
-                title: "Potassium",
-                value: store.food.potassium.measurement.converted(to: .milligrams),
-                breakdown: []
-            )
-        )
-    }
-
-    var sodium: some View {
-        NutritionalValueCard(
-            model: .init(
-                title: "Sodium",
-                value: store.food.sodium.measurement.converted(to: .milligrams),
-                breakdown: []
-            )
-        )
-    }
-
-    var energyBreakdown: some View {
-        GroupBox {
-            EnergyBreakdownChart(
-                food: store.food
-            )
-            .frame(height: 150)
-            .padding(.top)
-        } label: {
-            Text("Energy Breakdown")
-                .font(.title2)
-            Divider()
         }
     }
 
-    var quantityPicker: some View {
-        QuantityPicker(
-            store: store.scope(state: \.quantityPicker, action: \.quantityPicker)
-        )
-        .quantityPickerStyle(.dropdownGrouped)
+    @CasePathable
+    public enum Action {
+        case quantityPicker(QuantityPicker.Action)
     }
-}
 
-#Preview {
-    FoodDetails(
-        store: Store(
-            initialState: .init(
-                food: .preview
-            ),
-            reducer: {
-                FoodDetailsFeature()
+    public init() { }
+
+    public var body: some ReducerOf<Self> {
+        Scope(state: \.quantityPicker, action: \.quantityPicker) {
+            QuantityPicker()
+        }
+        Reduce { state, action in
+            switch action {
+                case .quantityPicker(let action):
+                    return reduce(state: &state, action: action)
             }
-        )
-    )
+        }
+    }
+
+    private func reduce(state: inout State, action: QuantityPicker.Action) -> Effect<Action> {
+        state.food = state.originalFood.changingServingSize(to: state.quantityPicker.quantity)
+        return .none
+    }
 }
