+++ food-spec/Sources/IngredientPicker/IngredientPicker.swift
@@ -1,90 +1,60 @@
-import SwiftUI
+import Foundation
 import Shared
 import QuantityPicker
 import ComposableArchitecture
 
-public struct IngredientPicker: View {
-    @Bindable var store: StoreOf<IngredientPickerFeature>
-    @State private var isDisclosed = false
+@Reducer
+public struct IngredientPicker {
+    @ObservableState
+    public struct State: Hashable {
+        public var food: Food
+        public var quantityPicker: QuantityPicker.State
+        public var isSelected: Bool
 
-    public init(store: StoreOf<IngredientPickerFeature>) {
-        self.store = store
-    }
+        public var ingredient: Ingredient {
+            .init(
+                food: food,
+                quantity: quantityPicker.quantity
+            )
+        }
 
-    public var body: some View {
-        DisclosureGroup(
-            isExpanded: self.$store.isSelected.sending(\.updateSelection)
+        public init(food: Food) {
+            self.food = food
+            self.quantityPicker = .init()
+            self.isSelected = false
+        }
+
+        public init(
+            food: Food,
+            quantity: Quantity
         ) {
-            QuantityPicker(
-                store: self.store.scope(
-                    state: \.quantityPicker,
-                    action: \.quantityPicker
-                )
-            )
-        } label: {
-            VStack(alignment: .leading, spacing: 8) {
-                Text(self.store.food.name.capitalized)
-                Text(self.store.ingredient.foodWithQuantity.nutritionalSummary)
-                    .font(.footnote)
-                    .foregroundStyle(.secondary)
+            self.food = food
+            self.quantityPicker = .init(quantity: quantity)
+            self.isSelected = true
+        }
+    }
+
+    @CasePathable
+    public enum Action {
+        case quantityPicker(QuantityPicker.Action)
+        case updateSelection(Bool)
+    }
+
+    public init() { }
+
+    public var body: some ReducerOf<Self> {
+        Scope(state: \.quantityPicker, action: \.quantityPicker) {
+            QuantityPicker()
+        }
+        Reduce { state, action in
+            switch action {
+                case .updateSelection(let selected):
+                    state.isSelected = selected
+                    return .none
+
+                case .quantityPicker:
+                    return .none
             }
         }
-        .disclosureGroupStyle(.selection)
     }
-}
-
-#Preview {
-    ScrollView {
-        ForEach(0..<5) { _ in
-            IngredientPicker(
-                store: .init(
-                    initialState: IngredientPickerFeature.State(food: .preview),
-                    reducer: {
-                        IngredientPickerFeature()
-                    }
-                )
-            )
-            .padding(.horizontal)
-        }
-    }
-}
-
-struct SelectionDisclosureGroupStyle: DisclosureGroupStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        GroupBox {
-            VStack(alignment: .leading) {
-                HStack(alignment: .top) {
-                    configuration.label
-                    Spacer()
-                    if configuration.isExpanded {
-                        Image(systemName: "checkmark.circle.fill")
-                            .foregroundStyle(.selection)
-                    } else {
-                        Image(systemName: "circle")
-                            .foregroundStyle(.secondary)
-                    }
-                }
-                .contentShape(UnevenRoundedRectangle(cornerRadii: .init(topLeading: 8, topTrailing: 8)))
-                .onTapGesture {
-                    configuration.$isExpanded.animation().wrappedValue.toggle()
-                }
-                configuration.content
-                    .padding(.top)
-                    .frame(height: configuration.isExpanded ? nil : 0)
-                    .clipped()
-                    .allowsHitTesting(configuration.isExpanded ? true : false)
-            }
-        }
-        .overlay {
-            RoundedRectangle(cornerRadius: 8)
-                .strokeBorder(
-                    configuration.isExpanded ? AnyShapeStyle(.selection) : AnyShapeStyle(.clear),
-                    lineWidth: 2
-                )
-        }
-    }
-}
-
-extension DisclosureGroupStyle where Self == SelectionDisclosureGroupStyle {
-    static var selection: Self { .init() }
 }
