+++ food-spec/Sources/MealList/MealList.swift
@@ -1,93 +1,112 @@
-import SwiftUI
+import Foundation
 import Shared
+import Database
 import MealForm
 import MealDetails
 import ComposableArchitecture
 
-public struct MealList: View {
-    @Bindable var store: StoreOf<MealListFeature>
+@Reducer
+public struct MealList {
+    @ObservableState
+    public struct State: Hashable {
+        var mealsWithNutritionalValues: [MealWithNutritionalValues] = []
+        @Presents var destination: Destination.State?
 
-    public init(store: StoreOf<MealListFeature>) {
-        self.store = store
+        var showsAddMealPrompt: Bool {
+            mealsWithNutritionalValues.isEmpty
+        }
+
+        struct MealWithNutritionalValues: Hashable {
+            let meal: Meal
+            let perTotal: Ingredient
+            let perServing: Ingredient
+        }
+
+        public init() { }
     }
 
-    public var body: some View {
-        List {
-            if store.showsAddMealPrompt {
-                ContentUnavailableView(
-                    "Your meals will be shown here.",
-                    systemImage: "fork.knife",
-                    description: Text("You can add a meal by tapping the \"+\" icon.")
-                )
-            } else {
-                mealsSection
-            }
-        }
-        .toolbar {
-            ToolbarItem(placement: .topBarTrailing) {
-                Button {
-                    store.send(.plusButtonTapped)
-                } label: {
-                    Image(systemName: "plus")
-                }
-            }
-        }
-        .navigationTitle("Meals")
-        .onFirstAppear {
-            store.send(.onFirstAppear)
-        }
-        .navigationDestination(
-            item: self.$store.scope(
-                state: \.destination?.mealDetails,
-                action: \.destination.mealDetails
-            ),
-            destination: { store in
-                MealDetails(store: store)
-            }
-        )
-        .sheet(
-            item: $store.scope(
-                state: \.destination?.mealForm,
-                action: \.destination.mealForm
-            ),
-            content: { store in
-                NavigationStack {
-                    MealForm(store: store)
-                }
-                .interactiveDismissDisabled()
-            }
-        )
+    @CasePathable
+    public enum Action {
+        case onFirstAppear
+        case plusButtonTapped
+        case onMealsUpdate([Meal])
+        case mealTapped(Meal)
+        case onDelete(IndexSet)
+        case destination(PresentationAction<Destination.Action>)
     }
 
-    private var mealsSection: some View {
-        Section {
-            ForEach(store.mealsWithNutritionalValues, id: \.meal.id) { nutritionalValue in
-                ListButton {
-                    self.store.send(.mealTapped(nutritionalValue.meal))
-                } label: {
-                    let footnotePerServingSize = "Per serving: \(nutritionalValue.perServing.foodWithQuantity.nutritionalSummary)"
-                    let footnotePerTotal = nutritionalValue.perTotal.foodWithQuantity.nutritionalSummary
-                    let footnote = nutritionalValue.meal.servings != 1 ? footnotePerServingSize : footnotePerTotal
-                    LabeledListRow(
-                        title: nutritionalValue.meal.name.capitalized,
-                        footnote: footnote
-                    )
-                }
+    public init() { }
+
+    @Dependency(\.databaseClient) private var databaseClient
+    @Dependency(\.nutritionalValuesCalculator) private var calculator
+
+    public var body: some ReducerOf<Self> {
+        Reduce { state, action in
+            switch action {
+                case .onFirstAppear:
+                    return .run { [databaseClient] send in
+                        let stream = databaseClient.observeMeals()
+                        for await meals in stream {
+                            await send(.onMealsUpdate(meals), animation: .default)
+                        }
+                    }
+
+                case .onMealsUpdate(let meals):
+                    state.mealsWithNutritionalValues = meals.map {
+                        .init(
+                            meal: $0,
+                            perTotal: calculator.nutritionalValues(meal: $0),
+                            perServing: calculator.nutritionalValuesPerServing(meal: $0)
+                        )
+                    }
+                    return .none
+
+                case .plusButtonTapped:
+                    state.destination = .mealForm(.init())
+                    return .none
+
+                case .mealTapped(let meal):
+                    state.destination = .mealDetails(.init(meal: meal))
+                    return .none
+
+                case .onDelete(let indices):
+                    return .run { [nutritionalValues = state.mealsWithNutritionalValues, databaseClient] send in
+                        let mealsToDelete = indices.map { nutritionalValues[$0].meal }
+                        for meal in mealsToDelete {
+                            try await databaseClient.delete(meal: meal)
+                        }
+                    }
+
+                case .destination:
+                    return .none
             }
-            .onDelete { offsets in
-                self.store.send(.onDelete(offsets))
+        }
+        .ifLet(\.$destination, action: \.destination) {
+            Destination()
+        }
+    }
+
+    @Reducer
+    public struct Destination {
+        @ObservableState
+        public enum State: Hashable {
+            case mealDetails(MealDetails.State)
+            case mealForm(MealForm.State)
+        }
+
+        @CasePathable
+        public enum Action {
+            case mealDetails(MealDetails.Action)
+            case mealForm(MealForm.Action)
+        }
+
+        public var body: some ReducerOf<Self> {
+            Scope(state: \.mealDetails, action: \.mealDetails) {
+                MealDetails()
+            }
+            Scope(state: \.mealForm, action: \.mealForm) {
+                MealForm()
             }
         }
     }
-}
-
-#Preview {
-    MealList(
-        store: .init(
-            initialState: MealListFeature.State(),
-            reducer: {
-                MealListFeature()
-            }
-        )
-    )
 }
