+++ food-spec/Tests/FoodListTests/BillboardReducerTests.swift
@@ -7,7 +7,7 @@ import ComposableArchitecture
 final class BillboardReducerTests: XCTestCase {
     func testOnAppear() async throws {
         let store = TestStore(
-            initialState: FoodListFeature.State(),
+            initialState: FoodList.State(),
             reducer: {
                 BillboardReducer()
             }
@@ -34,7 +34,7 @@ final class BillboardReducerTests: XCTestCase {
 
     func testOnAppear_billboardClientError() async throws {
         let store = TestStore(
-            initialState: FoodListFeature.State(),
+            initialState: FoodList.State(),
             reducer: {
                 BillboardReducer()
             }
@@ -50,7 +50,7 @@ final class BillboardReducerTests: XCTestCase {
 
     func testShowBanner() async throws {
         let store = TestStore(
-            initialState: FoodListFeature.State(),
+            initialState: FoodList.State(),
             reducer: {
                 BillboardReducer()
             }
